split measures into multiple files
and fix up math with measures
